+++ cli/factory.rs
@@ -12,8 +12,9 @@ use crate::cache::Caches;
 use crate::cache::DenoDir;
 use crate::cache::DenoDirProvider;
 use crate::cache::EmitCache;
+use crate::cache::GlobalHttpCache;
 use crate::cache::HttpCache;
-use crate::cache::HttpCachePaths;
+use crate::cache::LocalHttpCache;
 use crate::cache::NodeAnalysisCache;
 use crate::cache::ParsedSourceCache;
 use crate::emit::Emitter;
@@ -136,7 +137,8 @@ struct CliFactoryServices {
   deno_dir_provider: Deferred<Arc<DenoDirProvider>>,
   caches: Deferred<Arc<Caches>>,
   file_fetcher: Deferred<Arc<FileFetcher>>,
-  http_cache: Deferred<Arc<HttpCache>>,
+  global_http_cache: Deferred<Arc<GlobalHttpCache>>,
+  http_cache: Deferred<Arc<dyn HttpCache>>,
   http_client: Deferred<Arc<HttpClient>>,
   emit_cache: Deferred<EmitCache>,
   emitter: Deferred<Arc<Emitter>>,
@@ -235,12 +237,24 @@ impl CliFactory {
       .get_or_init(|| ProgressBar::new(ProgressBarStyle::TextOnly))
   }
 
-  pub fn http_cache(&self) -> Result<&Arc<HttpCache>, AnyError> {
+  pub fn global_http_cache(&self) -> Result<&Arc<GlobalHttpCache>, AnyError> {
+    self.services.global_http_cache.get_or_try_init(|| {
+      Ok(Arc::new(GlobalHttpCache::new(
+        self.deno_dir()?.deps_folder_path(),
+      )))
+    })
+  }
+
+  pub fn http_cache(&self) -> Result<&Arc<dyn HttpCache>, AnyError> {
     self.services.http_cache.get_or_try_init(|| {
-      Ok(Arc::new(HttpCache::new(HttpCachePaths {
-        global: self.deno_dir()?.deps_folder_path(),
-        local: self.options.remote_modules_dir_path(),
-      })))
+      let global_cache = self.global_http_cache()?.clone();
+      match self.options.remote_modules_dir_path() {
+        Some(local_path) => {
+          let local_cache = LocalHttpCache::new(local_path, global_cache);
+          Ok(Arc::new(local_cache))
+        }
+        None => Ok(global_cache),
+      }
     })
   }
 
@@ -530,6 +544,7 @@ impl CliFactory {
           self.maybe_file_watcher_reporter().clone(),
           self.emit_cache()?.clone(),
           self.file_fetcher()?.clone(),
+          self.global_http_cache()?.clone(),
           self.type_checker().await?.clone(),
         )))
       })
