+++ cli/tools/info.rs
@@ -95,7 +95,8 @@ fn print_cache_info(
   location: Option<&deno_core::url::Url>,
 ) -> Result<(), AnyError> {
   let dir = factory.deno_dir()?;
-  let modules_cache = factory.file_fetcher()?.get_http_cache_location();
+  #[allow(deprecated)]
+  let modules_cache = factory.global_http_cache()?.get_global_cache_location();
   let npm_cache = factory.npm_cache()?.as_readonly().get_cache_location();
   let typescript_cache = &dir.gen_cache.location;
   let registry_cache = dir.registries_folder_path();
